+++ VendinMachineApp/VendinMachineApp/Model/Beverages/Beverage.swift
